requirements.txt
polars==1.33.1
pandas==2.1.4
python-dotenv==1.1.1
connectorx==0.4.4
psycopg2-binary==2.9.10
sqlalchemy==1.4.54
gspread==6.2.1
google-auth==2.41.0
pyarrow==21.0.0
apache-airflow==3.0.0